fix: eliminate flaky 404s in link/QR live tests and fix qrCodeBytes type
The link integration tests intermittently failed with 'Fetch failed with
status 404'. Audit findings and fixes:

- The stats test listed the organization's short codes (newest first) and
  picked a random one. Concurrent CI runs (Node version matrix plus the
  coverage workflow) share the organization and delete their temporary
  codes within seconds, so the picked code often vanished before
  getCodeStats resolved. The test now creates and uses its own code.
- Operations that run immediately after a create can transiently 404
  while the API is eventually consistent. All live calls now go through a
  retry helper with backoff (replacing the one-off 500ms sleep), and list
  assertions poll until the created items are visible.
- Cleanup deletes were asserted, so a transient cleanup failure failed
  tests whose subject had already passed. Cleanup is now best-effort with
  a warning; dedicated delete tests assert the delete paths explicitly.

Also fixes qrCodeBytes to be Uint8Array: new Uint16Array(buffer) widened
each byte to 16 bits, which corrupts the QR image when the bytes are
written out.

+++ src/link.ts
@@ -109,7 +109,7 @@ export type CreateQrCodeResponse = {
 	id: string;
 	title?: string;
 	qrCode: string;
-	qrCodeBytes: Uint16Array;
+	qrCodeBytes: Uint8Array;
 	qrLink: string;
 };
 
@@ -538,7 +538,7 @@ export class Link extends BaseService {
 
 			if (result.qrCode) {
 				const buffer = Buffer.from(result.qrCode, "base64");
-				result.qrCodeBytes = new Uint16Array(buffer);
+				result.qrCodeBytes = new Uint8Array(buffer);
 			}
 
 			return result;
@@ -573,7 +573,7 @@ export class Link extends BaseService {
 
 			if (result.qrCode) {
 				const buffer = Buffer.from(result.qrCode, "base64");
-				result.qrCodeBytes = new Uint16Array(buffer);
+				result.qrCodeBytes = new Uint8Array(buffer);
 			}
 
 			return result;
@@ -612,7 +612,7 @@ export class Link extends BaseService {
 			for (const qrCode of result.data) {
 				if (qrCode.qrCode) {
 					const buffer = Buffer.from(qrCode.qrCode, "base64");
-					qrCode.qrCodeBytes = new Uint16Array(buffer);
+					qrCode.qrCodeBytes = new Uint8Array(buffer);
 				}
 			}
